scripts/requirements.txt
matplotlib>=3.8
pandas>=2.1
Pillow>=10.0
